+++ src/Header.css
@@ -33,6 +33,4 @@
     margin-right: 4%;
     color:#F3990F;
     outline: none;
-
-    
 }
